# tema4/ejem2a-p2-2023/src/main/java/es/codeurjc/booksmanagementspring/dto/BookCreateDTO.java
package es.codeurjc.booksmanagementspring.dto;

public record BookCreateDTO(
        String title,
        String resume,
        String author,
        String publishing,
        String year
) {
}