# tema4/ejem2a-p2-2023/src/main/java/es/codeurjc/booksmanagementspring/dto/BookDTO.java
package es.codeurjc.booksmanagementspring.dto;

import java.util.List;

public record BookDTO(
        long id,
        String title,
        String resume,
        String author,
        String publishing,
        String year,
        List<ReviewBasicDTO> reviews
)  {
}